requirements.txt
yfinance>=0.2.36
anthropic>=0.49.0
python-dotenv>=1.0.1
pandas>=2.2.0
